requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22
pandas>=1.5
scikit-learn>=1.1
matplotlib>=3.6
seaborn>=0.12
tqdm>=4.65
pillow>=9.2
grad-cam @ git+https://github.com/jacobgil/pytorch-grad-cam@master
torch
torchvision
tqdm
numpy
matplotlib
seaborn
scikit-learn
pillow